removed casting delta time to f32 every frame and just keeping f64 rather than f32

+++ devtest/src/game.cpp
@@ -8,16 +8,16 @@ b8 initialize(struct game* game_inst) {
 }
 
 // function pointer to game's update function
-b8 update(struct game* game_inst, f32 delta_time) {
+b8 update(struct game* game_inst, f64 delta_time) {
     return TRUE;
 }
 
 // function pointer to game's render function
-b8 render(struct game* game_inst, f32 delta_time) {
+b8 render(struct game* game_inst, f64 delta_time) {
     return TRUE;
 }
 
 // function pointer to handle resizes if applicable
-b8 on_resize(struct game* game_inst, f32 delta_time) {
+b8 on_resize(struct game* game_inst, f64 delta_time) {
     return TRUE;
 }
